Refactor Loader component’s name & its usage
