+++ src/components/core/Lightbox/styles.scss
